- move tests to MssqlCeQueryTests
